Update chunk loading
No longer render blocks at the edges of a chunk that are exposed.

Previously, had a merge method that would merge chunks after creating said chunks,

Instead, do that "merging" when creating a chunk by increasing the size of the chunk by 1 block on all sides of a chunk to get information about blocks in neighboring chunks, and figure out whether blocks on the edges of a chunk should be visible.

Now the application can scale with any minimum chunk height because it no longer changes the amount of data in the respective vertex buffer (e.g. previously the minimum height was -25 and increasing it would require a larger vertex buffer; however, now we can change it to any number, but keep in mind of the amount of memory from blocks, and we wouldn't need to change the vertex buffer size because it will be the same number of blocks; in this commit, we are able to set it to -500 with no problems with the vertex buffer).