Change dependency where player depends on world now
Previously, the world class depended on the player, so it always kept track of the player where the player has no notion of the world. A problem with this implementation arises with how blocks should be modified in the world if the player doesn't know about the world.

Now the player depends on the world; the player has a notion of the world while the world doesn't care what is in it (only cares about managing chunks).

This allows the player to modify the world as needed by asking the world to do things that need to be done, such as breaking and placing blocks.

Additionally, this implementation MAY allow for other entities to modify the world; thus, the world is affected by external influences. For example, in the future, we may have an enemy that destroys blocks. Maybe we have an explosive that affects the world. The world no longer needs to manage all of these entities and check for certain things; instead, these entities will let the world know manually. Futhermore, this should allow for greater flexibility and scalability.