Fix chunk multi-threading related to noise
Previously, there is a data race with setting the noise, where in one case, if one thread is interrupted while generating a chunk, another thread can reset the noise generator to default and immediately the original thread takes control again but now is using noise from the default configuration of the noise.

Now, we have multiple types of noises dependent on its responsibility, i.e., terrain heights. So, noise configurations are set independently and at the start of world creation.