Fix chunk bounds
Previously, the xz-bounds included +1 block than needed, i.e. a chunk of size 16 had bounds that spans 17 blocks and should be 16 blocks instead.

This caused issues with code that used the bounds as conditions. For example, when removing a block at a specific chunk edge, it reveals a neighboring block that doesn't belong to the chunk.

Fix by adding an offset when determining a chunk's bounds.