Change from 2D to 3D chunk loading
Previously, the program loaded chunks along the xz-plane with a height limit on the y-axis. This limited terrain generation in terms of height. Furthermore, each chunk potentially has a large memory footprint because a blocks of the chunk in the y-axis would need to load.

Instead, the program now accounts for the y-axis.

*Advantages:
- The program can now infinitely load chunks in terms of height, though memory-bounded.
- Less memory usage variation due to more consistent memory requirements for each chunk.

*Disadvantages:
- More overhead per chunk.

*Additional changes:
- Code cleanup.
- Update Model::Vertex hash function to account for normals.
- Color the blocks at the edges of a chunk red for debugging.
- Temporarily always keep vertex buffer filled with some kind of data.
- Add quad model, may use in future with instancing.