Update to allow empty vertex buffers
Suppose the player moves to an area where there are no blocks rendered, the respective vertex buffer would be empty.

In the previous implementation, updating a vertex buffer with an empty one would attempt to create an empty vertex buffer, which is not allowed.

Now, if the vertex buffer is empty, update the count so that the draw call doesn't draw any vertices; we ignore the data already present in the vertex buffer.