Update chunks to use less memory
Previously, each chunk stored at max 20*20*20 blocks, where 20 is derived from (16 blocks + 4 edge blocks), which uses quite a bit of memory unnecessarily as most of the blocks will be the same.

Now, we store a single instance of each block type and have chunks point to those blocks, kind of like a palette of blocks. This is possible due to chunks "storing" position data for the blocks already and that blocks don't need to know their position.
Since there is only 1 instance per block, attributes such as the collision shape of a block will be affected. One solution is to create a new collision shape related to the block when we want some physics (in this case, being able to break and place blocks). This is possible due to all blocks being the same at the moment. Hopefully, it won't be a problem in the future.

- Went from multiple instances of the same block to a single instance of a block to save memory.
- Update associated physics with the changes to block instancing.
- Update vertex shader; no longer shows size 16 chunk borders.
- Update "texture.cpp/hpp" to use VMA.
- Minor change for readability in "collision-handler.cpp".
- Some code clean up.