Refactor: shift to more fine grain chunk loading
Previously, the program regenerated the entire mesh for the world and sent all of that data to the GPU, which reduced performance in both memory and time.

Instead, we now create a new vertex and index buffer for each chunk. Everything time the program needs to render new chunks, remove them, or modify them, it can choose to do so with specific chunks instead of reloading the entire world.

Changes:
- Add a chunk loading system to "game.cpp/hpp" which handles concurrency; utilizes a dynamic mapping of chunk centers to vertex buffer ids.
- Modify chunk related callback functions in "world.cpp/hpp" where it will notify subscribers of either chunk additions/updates (load) or removal (unload).
- No longer generating the world in its constructor because we need to add callbacks before generating it; we now generate it explicitly with the method `init()`.
- Change working directory from project directory to directory containing the executables, so that when we run the executables, it is able to access the assets, e.g. textures.