Update asset management to avoid build generator dependence
Tested the program with Ninja and Visual Studio build generators; however, both placed the executable in different levels of directory.

To resolve, moved asset related into an assets folder and implemented a method that abstracts getting the assets folder and related assets. The implementation walks up from the executable directory until it reaches an assets folder and stores the path the that folder.

We can now requests asset paths, such as textures, using a method which abstracts away the need to know the path to the assets folder.

- Update respective locations in code that uses assets.
- Add utility method to get an asset from the assets folder.
- Update CMakeLists.txt to set startup project for Visual Studio, and re-enabled unix builds.
- Move shaders, models, and textures into an assets folder.
- Minor clean up.